BREAKING CHANGE - rename upside_down to upsidedown
