+++ index/upsidedown/analysis_test.go
@@ -1,4 +1,4 @@
-package upside_down
+package upsidedown
 
 import (
 	"testing"
