+++ index/upsidedown/benchmark_cznicb_test.go
@@ -9,7 +9,7 @@
 
 // +build cznicb
 
-package upside_down
+package upsidedown
 
 import (
 	"testing"
