+++ index/upsidedown/dump.go
@@ -7,7 +7,7 @@
 //  either express or implied. See the License for the specific language governing permissions
 //  and limitations under the License.
 
-package upside_down
+package upsidedown
 
 import (
 	"bytes"
