+++ index/upsidedown/upsidedown.go
@@ -7,7 +7,9 @@
 //  either express or implied. See the License for the specific language governing permissions
 //  and limitations under the License.
 
-package upside_down
+//go:generate protoc --gofast_out=. upsidedown.proto
+
+package upsidedown
 
 import (
 	"encoding/binary"
